Move registerTestExtensions to parser_test.go
